+++ profile_data/names.json
@@ -3,9 +3,11 @@
         "male": [
             "Alan",
             "Carl",
+            "Charles",
             "Donald",
             "Dylan",
             "Fernando",
+            "Francis",
             "Frederick",
             "Jacob",
             "Jake",
@@ -16,7 +18,9 @@
             "Michael",
             "Richard",
             "Robert",
+            "Roy",
             "Shaun",
+            "Thomas",
             "Troy",
             "Tyler",
             "William"
@@ -32,6 +36,7 @@
             "Helen",
             "Isabella",
             "Josephine",
+            "Joyce",
             "Mia",
             "Maya",
             "Meagan",
